+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <magtek_card_reader/magtek_card_reader_plugin_c_api.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  MagtekCardReaderPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("MagtekCardReaderPluginCApi"));
+}